+++ NEWS.md
@@ -1,5 +1,6 @@
 # gwasRtools 0.1.2
 
+* `get_loci()` "lead" column is the best lead SNP from distance (i.e., all input SNPs) or from LD clumping (subset in EUR panel)
 * Fixes for `get_loci()` niche cases
 
 # gwasRtools 0.1.1